fix: recalculate bounds after height adjustment
There was a issue with the linenumbers being rendered
wrongly when at the end of a file because the cursors
information about the height of the window to render
text was updated after adjusting the render viewport.

This fixed that issue but just moving the viewport
adjustment after the window height set in the cursor.