fix: offset bounds when search result is out of window bounds
This is a bit of a design issue, but ledit has no line wrap,
thats somewhat complex to implement with the current design.

The issue was that while searching x is the minibuffer position,
so it was only centering y, not x, as a fix when in search it
will offset the render window to the x needed by using the fallback value
as bound.
damn software design.